[TASK-2415] Reorganized classes
