+++ ServiceManagement/ServiceLocator.Generic.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: dcd0d9014cd29324899bff7e3097573b
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
