+++ ServiceManagement/ServiceLocator.cs.meta
@@ -1,11 +1,2 @@
 fileFormatVersion: 2
-guid: dcd0d9014cd29324899bff7e3097573b
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant: 
+guid: 38b304d29a2c58e4b9788b40283d50bc
